fix(clippy): drop redundant i64 cast on changedtick
buf_get_changedtick already returns i64, so `tick as i64` tripped
clippy::unnecessary_cast under -D warnings and broke CI. Return tick
directly.

+++ src/fusevm_bridge.rs
@@ -1436,7 +1436,7 @@ fn b_getvar(vm: &mut VM, _: u8) -> Value {
             .map_or(0, |b| {
                 crate::ported::buffer::buf_get_changedtick(&b.borrow())
             });
-        return Value::Int(tick as i64);
+        return Value::Int(tick);
     }
     // Dynamic v: state (not fixed v: constants).
     if name == "v:exception" {
